+++ CHANGELOG.md
@@ -10,6 +10,7 @@ refer to the issue number you resolved.
 
 - Features
   - Add frog-menu buffer regexps to `dimmer-configure-posframe`
+  - Automatically reset dimmer when the theme changes [#72]
 - Bugfixes
   - Handle `reset` face-attribute gracefully, resolving [#68, #69]
   - Handle faces that error on attribute lookup (e.g. `child-frame-border`), resolving [#53]